feat: refactor and better comment parser
