--- src/types/llm.ts
@@ -1,6 +1,0 @@
-export type TokenLimits = {
-  modelMaxTokenLimit: number;
-  maxCompletionTokens: number;
-  runningTokenCount: number;
-  tokensRemaining: number;
-};
